+++ lib/src/analyzers/lint_analyzer/rules/rules_factory.dart
@@ -30,6 +30,7 @@ import 'rules_list/ban_name/ban_name_rule.dart';
 import 'rules_list/binary_expression_operand_order/binary_expression_operand_order_rule.dart';
 import 'rules_list/check_for_equals_in_render_object_setters/check_for_equals_in_render_object_setters_rule.dart';
 import 'rules_list/component_annotation_arguments_ordering/component_annotation_arguments_ordering_rule.dart';
+import 'rules_list/consistent_update_render_object/consistent_update_render_object_rule.dart';
 import 'rules_list/double_literal_format/double_literal_format_rule.dart';
 import 'rules_list/format_comment/format_comment_rule.dart';
 import 'rules_list/member_ordering/member_ordering_rule.dart';
@@ -103,6 +104,7 @@ final _implementedRules = <String, Rule Function(Map<String, Object>)>{
       CheckForEqualsInRenderObjectSettersRule.new,
   ComponentAnnotationArgumentsOrderingRule.ruleId:
       ComponentAnnotationArgumentsOrderingRule.new,
+  ConsistentUpdateRenderObjectRule.ruleId: ConsistentUpdateRenderObjectRule.new,
   DoubleLiteralFormatRule.ruleId: DoubleLiteralFormatRule.new,
   FormatCommentRule.ruleId: FormatCommentRule.new,
   MemberOrderingRule.ruleId: MemberOrderingRule.new,